data/requirements.txt
airportsdata==20250224
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.6
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudscraper==1.2.71
cryptography==44.0.2
dotenv==0.9.9
emoji==2.14.1
exceptiongroup==1.2.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
kaitaistruct==0.10
numpy==2.0.2
outcome==1.3.0.post0
pandas==2.2.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==1.26.20
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zstandard==0.23.0
